withMotion: Handle animation on initial mount/render
This update adds a new prop/option to `withMotion` that can control the
animation rendering for a component. This prop is called
`isAnimateOnInitialMount`.

By default, it is set to `true`, which does not affect the animations for
current implementations.

If set to `false`, the component will only fire it's `componentDidMount`
animations after the initial page render.